BUG: No way to control date range for TableFilters component (#13886)
* BUG: No way to control date range for TableFilters component
Fixes #13885
