+++ frontend-react/package.json
@@ -67,10 +67,10 @@
         "build:production": "yarn run build-base",
         "build:csp": "yarn run build-base --mode csp",
         "build-base": "vite build",
-        "test": "cross-env TZ=UTC vitest",
-        "test:debug": "cross-env DEBUG_PRINT_LIMIT=100000 TZ=UTC vitest --run --no-file-parallelism",
-        "test:ci": "cross-env TZ=UTC VITE_BACKEND_URL=http://localhost vitest --coverage",
-        "test:ui": "cross-env TZ=UTC vitest --ui",
+        "test": "cross-env vitest",
+        "test:debug": "cross-env DEBUG_PRINT_LIMIT=100000 vitest --run --no-file-parallelism",
+        "test:ci": "cross-env VITE_BACKEND_URL=http://localhost vitest --coverage",
+        "test:ui": "cross-env vitest --ui",
         "test:e2e": "playwright test",
         "test:e2e-ui": "playwright test --ui",
         "lint": "eslint \"**/*.{js,ts,jsx,tsx}\" && prettier \"*\" --check --ignore-unknown && tsc",
