+++ frontend-react/src/components/DataDashboard/FacilitiesProviders/FacilitiesProvidersTable.test.tsx
@@ -8,7 +8,7 @@ import {
 } from "../../../__mocks__/OrganizationMockServer";
 import { FacilityResource } from "../../../config/endpoints/dataDashboard";
 import { mockSessionContentReturnValue } from "../../../contexts/__mocks__/SessionContext";
-import { mockFilterManager } from "../../../hooks/filters/mocks/MockFilterManager";
+import { filterManagerFixture } from "../../../hooks/filters/filters.fixtures";
 import { mockUseReceiverSubmitter } from "../../../hooks/network/DataDashboard/__mocks__/UseReceiverSubmitter";
 import { mockUseOrganizationReceivers } from "../../../hooks/network/Organizations/__mocks__/ReceiversHooks";
 import { renderApp } from "../../../utils/CustomRenderUtils";
@@ -40,6 +40,7 @@ const mockData: FacilityResource[] = [
 
 const mockReceivers = receiversGenerator(5);
 const mockActiveReceiver = mockReceivers[0];
+const mockFilterManager = { ...filterManagerFixture };
 
 vi.mock("rest-hooks", async (importActual) => ({
     ...(await importActual<typeof import("rest-hooks")>()),
