+++ frontend-react/src/hooks/filters/filters.fixtures.ts
@@ -1,15 +1,28 @@
 import { Dispatch } from "react";
 
-import { RangeSettingsAction } from "../UseDateRange";
-import { FilterManager } from "../UseFilterManager";
-import { PageSettingsAction } from "../UsePages";
-import { SortSettingsAction } from "../UseSortOrder";
+import { CursorManager } from "./UseCursorManager";
+import { RangeSettingsAction } from "./UseDateRange";
+import { FilterManager } from "./UseFilterManager";
+import { PageSettingsAction } from "./UsePages";
+import { SortSettingsAction } from "./UseSortOrder";
+
+export const cursorManagerFixture: CursorManager = {
+    cursors: {
+        current: "",
+        next: "",
+        history: [],
+    },
+    hasPrev: false,
+    hasNext: false,
+    // eslint-disable-next-line no-console
+    update: () => void 0,
+};
 
 const fakeDispatch = <T>(): Dispatch<T> => {
     return (_v: T) => void 0;
 };
 
-export const mockFilterManager: FilterManager = {
+export const filterManagerFixture: FilterManager = {
     rangeSettings: {
         from: new Date("2022-01-01").toISOString(),
         to: new Date("2022-12-31").toISOString(),
